+++ src/components/Button/mixins/buttonApi.ts
@@ -20,23 +20,23 @@ export function useMakeButtonProps(
 ) {
     return {
         class: {
-            'btn': !['icon', 'floating'].includes(<string>props.mode),
-            [`${cssPrefix}btn-${props.mode}`]: ['icon', 'floating'].includes(<string>props.mode),
+            "btn": !["icon", "floating"].includes(<string>props.mode),
+            [`${cssPrefix}btn-${props.mode}`]: ["icon", "floating"].includes(<string>props.mode),
             [`btn-outline-${props.color}`]: props.outlined && props.color && !props.transparent,
             [`btn-flat-${props.color}`]: !props.outlined && props.flat && props.color && !props.transparent,
             [`${cssPrefix}btn-transparent`]: !props.outlined && !props.flat && props.transparent,
             [`btn-${props.color}`]: !props.outlined && !props.flat && !props.transparent && props.color,
             [`${cssPrefix}btn-raised`]: props.raised,
             [`btn-${props.size}`]: !Helper.isEmpty(props.size),
-            'rounded-pill': props.mode !== 'icon' && props.pill,
-            'rounded-1': props.mode !== 'icon' && !props.pill && !props.rounded,
-            'disabled': disabled,
-            'active': props.active,
+            "rounded-pill": props.mode !== "icon" && props.pill,
+            "rounded-1": props.mode !== "icon" && !props.pill && !props.rounded,
+            "disabled": disabled,
+            "active": props.active,
         },
-        role: 'button',
-        type: buttonType === 'div' ? undefined : buttonType,
-        disabled: buttonType === 'div' ? undefined : disabled,
-        'aria-disabled': buttonType === 'div' ? undefined : disabled,
+        role: "button",
+        type: buttonType === "div" ? undefined : buttonType,
+        disabled: buttonType === "div" ? undefined : disabled,
+        "aria-disabled": buttonType === "div" ? undefined : disabled,
     }
 }
 
@@ -62,20 +62,16 @@ export function useMakeInputItemClasses(
     const isSelected = isInputItemSelected(item, props);
 
     return {
-        'btn': true,
+        "btn": true,
         [`btn-${props.toggleColor}`]: isSelected && props.toggleColor,
-        // [`btn-${props.color} active`]: isSelected && !props.toggleColor,
-        // [`btn-outline-${props.color}`]: !isSelected && props.outlined,
         [`btn-outline-${props.color}`]: props.outlined && !props.toggleColor,
-        // [`btn-flat-${props.color}`]: !isSelected && !props.outlined && props.flat,
         [`btn-flat-${props.color}`]: props.flat && !props.outlined && !props.toggleColor,
-        // [`btn-${props.color}`]: !isSelected && !props.outlined && !props.flat,
         [`btn-${props.color}`]: !props.outlined && !props.flat && !props.toggleColor,
         [`btn-${props.size}`]: props.size,
         [`${cssPrefix}btn-raised`]: props.raised,
-        'active': isSelected && !props.toggleColor && !props.disabled && !item.disabled,
-        'disabled': props.disabled || item.disabled,
-        'readonly': props.readonly || item.readonly,
+        "active": isSelected && !props.toggleColor && !props.disabled && !item.disabled,
+        "disabled": props.disabled || item.disabled,
+        "readonly": props.readonly || item.readonly,
     }
 }
 
@@ -85,21 +81,21 @@ export function useMakeInputItemAttrs(
 ): TRecord | object {
     const attr = {
         id: item.id,
-        role: props.multiple ? 'checkbox' : 'radio',
-        type: props.multiple ? 'checkbox' : 'radio',
+        role: props.multiple ? "checkbox" : "radio",
+        type: props.multiple ? "checkbox" : "radio",
         name: props.multiple ? item.name : props.name,
         required: props.required,
         disabled: props.disabled || item.disabled,
         readonly: props.readonly || item.readonly,
-        'aria-disabled': props.disabled || item.disabled,
-        'aria-checked': isInputItemSelected(item, props),
+        "aria-disabled": props.disabled || item.disabled,
+        "aria-checked": isInputItemSelected(item, props),
     };
 
     if (props.multiple) {
         return {
             ...attr,
-            'true-value': true,
-            'false-value': false
+            "true-value": true,
+            "false-value": false
         }
     }
 
@@ -119,7 +115,7 @@ export function useRenderIconWithSlot(
         return h("span", {
                 class: {
                     [`${cssPrefix}icon`]: true,
-                    [`${cssPrefix}icon-${iconPosition}`]: btnMode !== 'icon' && btnMode !== 'floating',
+                    [`${cssPrefix}icon-${iconPosition}`]: btnMode !== "icon" && btnMode !== "floating",
                 }
             },
             // @ts-ignore
@@ -132,7 +128,7 @@ export function useRenderIconWithSlot(
             !Helper.isEmpty(props.icon)
                 ? h<TBsIcon>(BsIcon, {
                     class: {
-                        [`${cssPrefix}icon-${iconPosition}`]: btnMode !== 'icon' && btnMode !== 'floating',
+                        [`${cssPrefix}icon-${iconPosition}`]: btnMode !== "icon" && btnMode !== "floating",
                     },
                     size: <Prop<string | number>>iconSize,
                     ...useCreateIconProps(props),
@@ -147,25 +143,25 @@ export function useRenderButtonContent(
     props: Readonly<TButtonOptionProps>,
 ): VNodeArrayChildren {
     return [
-        (props.iconPosition === 'left')
+        (props.iconPosition === "left")
             ? useRenderIconWithSlot(
-                slots, 'icon',
+                slots, "icon",
                 <string>props.mode,
                 props,
                 props.iconPosition,
                 <number>props.iconSize,
             )
-            : '',
+            : "",
         slots.default && slots.default(),
-        (props.iconPosition === 'right')
+        (props.iconPosition === "right")
             ? useRenderIconWithSlot(
-                slots, 'icon',
+                slots, "icon",
                 <string>props.mode,
                 props,
                 props.iconPosition,
                 <number>props.iconSize,
             )
-            : '',
+            : "",
     ]
 }
 
@@ -176,34 +172,34 @@ export function useRenderToggleItemContent(
     // inputField: VNode,
 ): VNodeArrayChildren {
     return [
-        (props.iconPosition === 'left')
+        (props.iconPosition === "left")
             ? useRenderIconWithSlot(
-                slots, 'icon',
-                'default',
+                slots, "icon",
+                "default",
                 item,
                 props.iconPosition,
                 item.iconSize,
                 item,
             )
-            : '',
+            : "",
         // inputField,
         useRenderSlot(
-            slots, 'default',
+            slots, "default",
             {key: Helper.uuid()},
-            h('span', {
+            h("span", {
                 class: `${cssPrefix}btn-text`,
             }, item.label),
             item,
         ),
-        (props.iconPosition === 'right')
+        (props.iconPosition === "right")
             ? useRenderIconWithSlot(
-                slots, 'icon',
-                'default',
+                slots, "icon",
+                "default",
                 item,
                 props.iconPosition,
                 item.iconSize,
                 item,
             )
-            : '',
+            : "",
     ]
 }