+++ src/components/Chip/mixins/chipApi.ts
@@ -19,8 +19,8 @@ export function useChipClassNames(
 ): TRecord {
     return {
         [`${cssPrefix}chip`]: true,
-        [`${cssPrefix}chip-sm`]: props.size === 'sm',
-        [`${cssPrefix}chip-lg`]: props.size === 'lg',
+        [`${cssPrefix}chip-sm`]: props.size === "sm",
+        [`${cssPrefix}chip-lg`]: props.size === "lg",
         [`${cssPrefix}chip-pill`]: props.pill,
         [`${cssPrefix}chip-clickable`]: (
             (props.href || attrs.click || attrs.onclick || attrs.onClick) && !props.disabled
@@ -28,8 +28,8 @@ export function useChipClassNames(
         [`${cssPrefix}chip-${props.color}`]: props.color && !props.outlined && (!props.activeClass || props.active === false),
         [`${cssPrefix}chip-outline-${props.color}`]: props.color && props.outlined && (!props.activeClass || props.active === false),
         [<string>props.activeClass]: props.activeClass && (props.active === true) && !props.disabled,
-        'active': (props.active === true) && !props.disabled && !props.activeClass,
-        'disabled': props.disabled === true,
+        "active": (props.active === true) && !props.disabled && !props.activeClass,
+        "disabled": props.disabled === true,
     }
 }
 
@@ -39,12 +39,12 @@ function getChipAvatarSize(
 ) {
     let imgSize: string;
 
-    if (chipSize === 'sm') {
-        imgSize = hasPadding ? '1.125rem' : '1.56rem';
-    } else if (chipSize === 'lg') {
-        imgSize = hasPadding ? '2.375rem' : '3rem';
+    if (chipSize === "sm") {
+        imgSize = hasPadding ? "1.125rem" : "1.56rem";
+    } else if (chipSize === "lg") {
+        imgSize = hasPadding ? "2.375rem" : "3rem";
     } else {
-        imgSize = hasPadding ? '1.5rem' : '2rem';
+        imgSize = hasPadding ? "1.5rem" : "2rem";
     }
 
     return {
@@ -57,14 +57,14 @@ function renderChipAvatar(props: Readonly<TChipOptionProps>): VNode {
     return h("div", {
         class: [
             `${cssPrefix}chip-avatar`,
-            props.imgPadding === false ? `${cssPrefix}chip-avatar-bounded` : '',
+            props.imgPadding === false ? `${cssPrefix}chip-avatar-bounded` : "",
         ]
     }, [
         h("img", {
             src: props.imgSrc,
-            alt: 'Chip Avatar',
+            alt: "Chip Avatar",
             class: [
-                props.imgCircle || props.pill ? 'rounded-circle' : 'rounded'
+                props.imgCircle || props.pill ? "rounded-circle" : "rounded"
             ],
             style: getChipAvatarSize(props.size, props.imgPadding)
         })
@@ -78,14 +78,14 @@ function createCloseBtnAttr(
     return <TBsButton>{
         // @ts-ignore
         flat: true as Prop<boolean>,
-        mode: 'icon' as Prop<string>,
-        icon: 'close' as Prop<string>,
+        mode: "icon" as Prop<string>,
+        icon: "close" as Prop<string>,
         iconSize: <Prop<number>>(
-            props.size === 'sm' ? 14 : props.size === 'lg' ? 22 : 20
+            props.size === "sm" ? 14 : props.size === "lg" ? 22 : 20
         ),
-        size: <Prop<string | undefined>>(props.size === 'sm' ? 'xs' : 'sm'),
-        color: ['light', 'light-grey'].includes(<string>props.color) ? 'dark' : props.color,
-        onClick: (): void => clickHandler(),
+        size: <Prop<string | undefined>>(props.size === "sm" ? "xs" : "sm"),
+        color: ["light", "light-grey"].includes(<string>props.color) ? "dark" : props.color,
+        onClick: clickHandler,
     }
 }
 
@@ -109,16 +109,16 @@ export function useRenderChip(
         }, {
             default: () => [
                 useRenderTransition(
-                    {name: 'scale'},
+                    {name: "scale"},
                     useRenderSlotWithWrapper(
-                        slots, 'chipIcon', Helper.uuid(), "div",
+                        slots, "chipIcon", Helper.uuid(), "div",
                         {class: `${cssPrefix}chip-icon`},
                         !Helper.isEmpty(props.icon)
                             ? h<TBsIcon>(BsIcon, {
                                 ...useCreateIconProps(props),
                                 icon: <Prop<string>>(`${props.icon}_${props.iconVariant}`),
                                 size: <Prop<string | number>>(
-                                    props.size === 'sm' ? 18 : (props.size === 'lg' ? 40 : 24)
+                                    props.size === "sm" ? 18 : (props.size === "lg" ? 40 : 24)
                                 ),
                             })
                             : undefined,